Major dtype bug caused all blank causal steering and potential biases in KL and latent/bary

+++ validation/top_tokens_by_vertex.py
@@ -333,7 +333,7 @@ def main():
         args.model_name,
         device=args.device,
         cache_dir=args.cache_dir,
-        dtype=torch.float16,
+        dtype=torch.bfloat16,
     )
     tokenizer = model.tokenizer
     print(f"  Model loaded. Vocab size: {model.cfg.d_vocab}")
